requirements.txt
streamlit
tensorflow
numpy
pillow # Pillow is required for PIL.Image

print("requirements.txt created successfully.")
